requirements.txt
requests
BeautifulSoup4
customtkinter
pillow
psycopg2
matplotlib
